выполнить 4-й шаг

+++ src/main/java/hexlet/code/controller/UserController.java
@@ -3,7 +3,9 @@ package hexlet.code.controller;
 import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserDTO;
 import hexlet.code.dto.UserUpdateDTO;
+import hexlet.code.exception.AccessDeniedException;
 import hexlet.code.service.UserService;
+import hexlet.code.util.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -26,6 +28,9 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserUtils userUtils;
+
     @GetMapping("")
     public List<UserDTO> index() {
         return userService.getAll();
@@ -44,12 +49,18 @@ public class UserController {
 
     @PutMapping("/{id}")
     public UserDTO update(@PathVariable Long id, @Valid @RequestBody UserUpdateDTO data) {
+        if (userUtils.getCurrentUser().getId() != id) {
+            throw new AccessDeniedException("You do not have permission to update this user");
+        }
         return userService.update(id, data);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
+        if (userUtils.getCurrentUser().getId() != id) {
+            throw new AccessDeniedException("You do not have permission to delete this user");
+        }
         userService.delete(id);
     }
 }
